Confirmation, toast and minor bugs

+++ src/app/registered/booklists-details/booklists-details.component.ts
@@ -5,6 +5,7 @@ import { Booklist } from '../../interface/booklist';
 import { UiService } from 'src/app/ui/ui.service';
 import { CategoryService } from '../../service/category.service';
 import { Category } from '../../interface/category';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-booklists-details',
@@ -49,14 +50,40 @@ export class BooklistsDetailsComponent implements OnInit {
   }
 
   handleDelete(booklist: Booklist) {
-    this.booklistService.delete(booklist.id).subscribe(
-      () => {
-        this.router.navigateByUrl('/profil');
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-secondary ml-1',
+        cancelButton: 'btn btn-warning mr-1',
       },
-      (error) => {
-        this.error =
-          'Une erreur semble être survenue lors de la suppression de la booklist. Veuillez nous excusez pour le désagréement.';
-      }
-    );
+      buttonsStyling: false,
+    });
+
+    swalWithBootstrapButtons
+      .fire({
+        title: 'Etes-vous sûre de vouloir supprimer cette booklist ?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Oui, vas-y',
+        cancelButtonText: 'Non, annule',
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.value) {
+          this.booklistService.delete(booklist.id).subscribe(
+            () => {
+              swalWithBootstrapButtons.fire(
+                'Supppression',
+                'Votre booklist a bien été supprimée.',
+                'success'
+              );
+              this.router.navigateByUrl('/profil');
+            },
+            (error) => {
+              this.error =
+                'Une erreur semble être survenue lors de la suppression de la booklist. Veuillez nous excusez pour le désagréement.';
+            }
+          );
+        }
+      });
   }
 }
